+++ Sources/ConControlsTests/UnitTests/Controls/TextControl/Text.cs
@@ -23,11 +23,11 @@ namespace ConControlsTests.UnitTests.Controls.TextControl
             sut.Invoking(s => s.Text = null!).Should().Throw<ArgumentNullException>();
         }
         [TestMethod]
-        public void Text_NotNull_SetInControllerAndEventRaised()
+        public void Text_NotNull_SetInControllerEventRaisedAndDrawn()
         {
             var stubbedWindow = new StubbedWindow();
             var textController = new StubbedConsoleTextController();
-            var sut = new StubbedTextControl(stubbedWindow, textController);
+            var sut = new StubbedTextControl(stubbedWindow, textController) {Parent = stubbedWindow};
 
             const string expectedText = "expected";
             string controllerValue = string.Empty;
@@ -36,16 +36,22 @@ namespace ConControlsTests.UnitTests.Controls.TextControl
             int eventRaised = 0;
             sut.TextChanged += (sender, e) => eventRaised += 1;
 
+            bool drawn = false;
+            stubbedWindow.Graphics.CopyCharactersConsoleColorConsoleColorPointCharArraySize = (color, consoleColor, arg3, arg4, arg5) => drawn = true;
+
             sut.Text = expectedText;
+            drawn.Should().BeTrue();
             controllerValue.Should().Be(expectedText);
             sut.Text.Should().Be(expectedText);
             eventRaised.Should().Be(1);
             
             // don't raise event twice
+            drawn = false;
             sut.Text = expectedText;
             controllerValue.Should().Be(expectedText);
             sut.Text.Should().Be(expectedText);
             eventRaised.Should().Be(1);
+            drawn.Should().BeFalse();
         }
     }
 }